requirements.txt
asgiref==3.8.1
dj-database-url==1.0.0
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
Pillow==9.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2022.4
